[S10P12A306-658] refactor: Modify BoardResponse

+++ BackEnd/bid/bid-student/src/main/java/com/ssafy/bid/domain/board/dto/BiddingCreateRequest.java
@@ -15,7 +15,7 @@ public class BiddingCreateRequest {
 	private int gradeNo;
 	private long boardNo;
 
-	private BiddingCreateRequest(int price, int userNo, int gradeNo, int boardNo) {
+	private BiddingCreateRequest(int price, int userNo, int gradeNo, long boardNo) {
 		this.price = price;
 		this.userNo = userNo;
 		this.gradeNo = gradeNo;
